+++ src/apps/synthetix/common/synthetix.position-presenter.ts
@@ -1,5 +1,6 @@
 import { Inject } from '@nestjs/common';
 import { formatBytes32String } from 'ethers/lib/utils';
+import _ from 'lodash';
 
 import { Register } from '~app-toolkit/decorators';
 import {
@@ -7,7 +8,7 @@ import {
   buildPercentageDisplayItem,
   buildNumberDisplayItem,
 } from '~app-toolkit/helpers/presentation/display-item.present';
-import { ContractPosition } from '~position/position.interface';
+import { ContractType } from '~position/contract.interface';
 import { PositionPresenterTemplate, ReadonlyBalances } from '~position/template/position-presenter.template';
 
 import { SynthetixContractFactory } from '../contracts';
@@ -21,10 +22,29 @@ export abstract class SynthetixPositionPresenter extends PositionPresenterTempla
 
   @Register.BalanceProductMeta('Mintr')
   async getMintrMeta(address: string, balances: ReadonlyBalances) {
-    const position = balances[0] as ContractPosition;
-    const snxBalance = position.tokens?.find(v => v.symbol === 'SNX');
-    const sUSDBalance = position.tokens?.find(v => v.symbol === 'sUSD');
-    if (!snxBalance || !sUSDBalance) return [];
+    let snxPrice: number | undefined;
+    let susdPrice: number | undefined;
+
+    // Search for the first position with the SNX & SUSD token to get the price of the 2 tokens.
+    for (const b of balances) {
+      if (b.type !== ContractType.POSITION) continue;
+
+      const tokensBySymbols = _(b.tokens)
+        .groupBy(t => t.symbol)
+        .mapValues(v => v[0])
+        .value();
+
+      const snxToken = tokensBySymbols['SNX'];
+      const susdToken = tokensBySymbols['sUSD'];
+
+      if (snxToken && susdToken) {
+        snxPrice = snxToken.price;
+        susdPrice = susdToken.price;
+        break;
+      }
+    }
+
+    if (!snxPrice || !susdPrice) return [];
 
     const synthetixContract = this.contractFactory.synthetixNetworkToken({
       address: this.snxAddress,
@@ -41,8 +61,8 @@ export abstract class SynthetixPositionPresenter extends PositionPresenterTempla
     const collateralBalance = Number(collateralRaw) / 10 ** 18;
     const unlockedSnx = Number(unlockedSnxRaw) / 10 ** 18;
     const debtBalance = Number(debtBalanceRaw) / 10 ** 18;
-    const collateralUSD = collateralBalance * snxBalance.price;
-    const debtBalanceUSD = -debtBalance * sUSDBalance.price;
+    const collateralUSD = collateralBalance * snxPrice;
+    const debtBalanceUSD = -debtBalance * susdPrice;
     const cRatio = debtBalance > 0 ? (collateralUSD / debtBalance) * 100 : 1;
     const escrowed = collateralBalance - unlockedSnx;
     const unescrowed = unlockedSnx;
@@ -53,7 +73,7 @@ export abstract class SynthetixPositionPresenter extends PositionPresenterTempla
       { label: 'C-Ratio', ...buildPercentageDisplayItem(cRatio) },
       { label: 'Escrowed SNX', ...buildNumberDisplayItem(escrowed) },
       { label: 'Unescrowed SNX', ...buildNumberDisplayItem(unescrowed) },
-      { label: 'SNX Price', ...buildDollarDisplayItem(snxBalance.price) },
+      { label: 'SNX Price', ...buildDollarDisplayItem(snxPrice) },
     ];
   }
 }